requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
astroid==3.1.0
bcrypt==4.1.2
bidict==0.23.1
black==24.3.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
fastapi==0.110.1
filelock==3.13.3
flake8==7.0.0
fpdf==1.7.2
future==1.0.0
greenlet==3.0.3
h11==0.14.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
iso8601==2.1.0
isort==5.13.2
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
nodeenv==1.8.0
packaging==24.0
paho-mqtt==2.0.0
pathspec==0.12.1
pdfkit==1.0.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
psycopg2==2.9.9
py-dotenv==0.1
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.1.0
pytest==8.1.1
python-engineio==4.9.0
PyYAML==6.0.1
requests==2.31.0
serial==0.0.97
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomlkit==0.12.4
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
websocket-client==1.7.0
websockets==12.0
wsproto==1.2.0
